Added a method for print the minimum and maximum order. Changed the data type for the order amount in the order class.

+++ Entities/Order.cs
@@ -17,5 +17,5 @@ public class Order
     public DateTime? OrderDate { get; set; }
 
     [Column("total_amount")]
-    public int TotalAmount { get; set; }
+    public decimal TotalAmount { get; set; }
 }
